moving FantasizedGPRModel
